+++ samples/react-native/babel.config.js
@@ -1,5 +1,3 @@
-const componentAnnotatePlugin = require('@sentry/babel-plugin-component-annotate');
-
 module.exports = {
   presets: ['module:@react-native/babel-preset'],
   plugins: [
@@ -11,6 +9,6 @@ module.exports = {
         },
       },
     ],
-    componentAnnotatePlugin,
+    'react-native-reanimated/plugin',
   ],
 };