+++ src/js/client.ts
@@ -18,6 +18,8 @@ import { Alert } from 'react-native';
 import { createIntegration } from './integrations/factory';
 import { defaultSdkInfo } from './integrations/sdkinfo';
 import type { ReactNativeClientOptions } from './options';
+import type { mobileReplayIntegration } from './replay/mobilereplay';
+import { MOBILE_REPLAY_INTEGRATION_NAME } from './replay/mobilereplay';
 import { ReactNativeTracing } from './tracing';
 import { createUserFeedbackEnvelope, items } from './utils/envelope';
 import { ignoreRequireCycleLogs } from './utils/ignorerequirecyclelogs';
@@ -44,7 +46,6 @@ export class ReactNativeClient extends BaseClient<ReactNativeClientOptions> {
     super(options);
 
     this._outcomesBuffer = [];
-    this._initNativeSdk();
   }
 
   /**
@@ -107,6 +108,14 @@ export class ReactNativeClient extends BaseClient<ReactNativeClientOptions> {
   }
 
   /**
+   * @inheritDoc
+   */
+  public init(): void {
+    super.init();
+    this._initNativeSdk();
+  }
+
+  /**
    * Sets up the integrations
    */
   protected _setupIntegrations(): void {
@@ -160,7 +169,14 @@ export class ReactNativeClient extends BaseClient<ReactNativeClientOptions> {
    * Starts native client with dsn and options
    */
   private _initNativeSdk(): void {
-    NATIVE.initNativeSdk(this._options)
+    NATIVE.initNativeSdk({
+      ...this._options,
+      mobileReplayOptions:
+        this._integrations[MOBILE_REPLAY_INTEGRATION_NAME] &&
+        'options' in this._integrations[MOBILE_REPLAY_INTEGRATION_NAME]
+          ? (this._integrations[MOBILE_REPLAY_INTEGRATION_NAME] as ReturnType<typeof mobileReplayIntegration>).options
+          : undefined,
+    })
       .then(
         (result: boolean) => {
           return result;